Fix database health check test (#1)
We copied the existing `test/services/database_health_check.service.test.js` directly from [sroc-charging-module-api](https://github.com/DEFRA/sroc-charging-module-api). In that project we expect 5 key tables to exist. Or, we did at the time we wrote the test. We would now expect more!

When we copied the test across it all worked fine because we were again pointing at an existing DB with lots of tables. However, when this project got pushed to GitHub the CI started failing because we were connecting to a DB that has nothing!

We've yet to add any migrations so nothing exists but the DB. We've confirmed that as long as the DB exists the connection works and that is sufficient as a health check.

Rather than force a result from PostgreSQL we're happy the fact we don't error is confirmation that the DB is 'healthy'.