Add .vscode to .gitignore
We maintain some VSCode `tasks.json` and `launch.json` in our main [Vagrant](https://www.vagrantup.com/) repo. This is held on a private GitLab instance for use only by the dev team.

These get copied into the projects to help the team work with the Vagrant box. They are specific to the box so make no sense if you are not using it.

Because of this we don't include them in source control. But when we created this project we used [sroc-charging-module-api](https://github.com/DEFRA/sroc-charging-module-api) as a template and it doesn't include `.vscode` in its `.gitignore` file.

This adds the entry, brings this prject inline with the rest of the other WABS projects, and ensures our vscode files don't get comitted.