Add proxyrequire and stub exec() calls
This was a pig of a problem! We are trying to stub `exec()` that is exported from the module `child_process` and expects to be passed a callback. It is then promisified by `util.promisify()` before being used in the service. We also want it to return 2 different values to mimic it being used to check ClamAV and Redis.

We tried various suggestions for how to do this, but none worked until we brought in [proxyquire](https://github.com/thlorenz/proxyquire)

> Proxies nodejs's require in order to make overriding dependencies during testing easy while staying totally unobtrusive.

With this change, and a minor one in the service code, we are now mocking all calls to external services, whether that be other web services or the shell.