Add tests for when services are down
We wanted to cover what happens if when making a call to one of the other services we get an error. Prior to this change if that happened the `ServiceStatusService` would implode and cause a `500` to be returned to the browser.

We need the service to handle this so it can continue to report what's going on with the other services. So, we added some tests that mock different scenarios. This meant some changes in the test setups, most notably having to move some of the Nock calls down into other blocks.

But to make it work in the code required some further refactoring. Now all the requests are made in one method, mainly to avoid duplicating the error handling all other the place. We've also brought across our experience of using Got in sroc-charging-module-api and use request options to control when exactly it should retry requests.

Finally, with things coming together we add some more notes about the test setup.