Thoughts on current state
Looking at the tests we see a copious the amount of work needed to mock web requests and system calls. We also have a range of scenarios. IMHO they are telling me that our service is doing too much. It is collecting data from different sources in different ways, which means it definately doesn't just have a single responsiblity.

It needs breaking up, which in turn means we can break up the tests and hopefully make them a damn sight less scary.

But GitHub currently clocks the PR at +905/-125. That's more than a 1000 changes and well beyond our aim of approzimately 100. We also need to get onto delivering some features. So, for now it works and is fully tested. Pending any quick wins we spot, we should approve and merge this and return to it at a later date (would be a great refactoring exercise for those of want the practise!)