Replace node-sass with dart-sass package (#14)
The `node-sass` package has been added as a dev dependency but is required as part of the build. It therefore needs to be moved to the dependencies section of the `package.json`

We realised while doing this that node-sass has loads of vulnerabilities and is deprecated so we've replaced it with a shiny new package dart-sass as per the docs for the gov.uk frontend.