Rename test config
We need somewhere to put our config for whether to include asset requests in the log. We already have a setting to control whether the log should record anything when running in test. So, putting these together makes the most sense.

Having done then, we concluded this isn't really test config, but log config hence we have renamed the file.