Filter asset requests from Hapi-pino logs (#16)
Our current hapi-pino configuration is based on what we needed for the [sroc-charging-module-api](https://github.com/DEFRA/sroc-charging-module-api). But as an API it never had to worry about assets and requests from the browser for them.

As we use this app, we're seeing multiple log entries per page. This is because it's logging both the main request plus one for each asset used by the page. Below is an example of what we see. Bear in mind that we have only listed the top-level entry. Each one is 41 lines long (when working locally) because the logger details both the request and response.

```text
[09:41:06.763] INFO (17542): [response] get /service-status 200 (1138ms)
[09:41:06.871] INFO (17542): [response] get /assets/stylesheets/application.css 200 (72ms)
[09:41:06.873] INFO (17542): [response] get /assets/all.js 200 (64ms)
[09:41:06.893] INFO (17542): [response] get /assets/images/govuk-crest.png 200 (8ms)
[09:41:06.926] INFO (17542): [response] get /assets/fonts/light-94a07e06a1-v2.woff2 200 (19ms)
[09:41:06.928] INFO (17542): [response] get /assets/fonts/bold-b542beb274-v2.woff2 200 (18ms)
[09:41:07.032] INFO (17542): [response] get /assets/images/favicon.ico 200 (6ms)
```

Our primary reason for using hapi-pino is to reduce the noise in our logs so we can find and focus on what's important. This change updates the hapi-pino filtering so we only log the main requests. It will include an option to switch back on asset logging, just in case we ever have issues and need that extra detail.

** Notes

- Rename test config

We needed somewhere to put our config for whether to include asset requests in the log. We already have a setting to control whether the log should record anything when running in `test`. So, putting these together made the most sense.

Having done that, we concluded this wasn't really test config anymore, but log config. Hence we renamed the file.