Move location where 'dotenv' is being required (#21)
In a previous commit https://github.com/DEFRA/water-abstraction-system/pull/10 the 'dotenv' package was moved as early as possible in the application.

Unfortunately this has turned out to be a bit too early and is causing some issues when running the unit tests locally. The unit tests don't use the /index.js so 'dotenv' is never required.

The 'dotenv' require therefore needs to be moved up a level to /app/server.js.