Create test database migrations (#22)
https://eaflood.atlassian.net/browse/WATER-3787

Some migrations scripts are required to create the tables/data for the unit tests in this PR https://github.com/DEFRA/water-abstraction-system/pull/15 in order for the build process to pass.

As no migrations process had yet been set up on this application. This PR has been created to build the migrations process as well as populate the test database with the data required for the PR mentioned above.

This migration process has been based on that used by the [Charging Module](https://github.com/DEFRA/sroc-charging-module-api).