Add timeout option to requests
When we spotted the issue with requests failing when behind a proxy, one of the behaviours we saw was that it took _forever_ to actually raise the exception.

This is because Got by default won't timeout attempts. If a request has not completed in 5 secs we're happy to accept that its failed.

So, we have updated our new service to support specifying a timeout of 5 seconds for all Got requests. This meant adding some new config, which includes proxy config that we're gonna look at in the next requests.