Select SROC supplementary charge versions for real
https://eaflood.atlassian.net/browse/WATER-3787

In [Fetching the charge versions data](https://github.com/DEFRA/water-abstraction-system/pull/15) we added a new service to start selecting scharge versions to be included in the Supplementary Bill run.

But at the moment, it's only filter is all charge versions with the `scheme` SROC. We have test data and scenarios to work with so we can now start implementing the filter 'for real'.

This is the start of the process. We expect it's likely we'll be finessing the filter in subsequent PR's. But this gets the ball rolling.