Update knex query in SupplementaryService
Two changes here. First me re-ordered the method calls and swapped some out for other aliases in order for the query to read more like traditional SQL.

The second is that we have updated it to take into account the `include_in_supplementary_billing` flag on the licences table. This is the main driver of whether to include something in supplementary billing.

“Simple”, was our next thought. Charge versions have a `scheme` and a `status` column; we'll just grab the one where `scheme=sroc && status=current` for each licence flagged for supplementary billing. That was until our initial tests showed that `status` is not getting set correctly. This means we have multiple SROC charge versions both with a status of `current`.

Fortunately, it looks like `end_date` is getting set as expected, which means we have a way of selecting them (for now!)