Select SROC supplementary charge versions for real (#27)
https://eaflood.atlassian.net/browse/WATER-3787

In [Fetching the charge versions data](https://github.com/DEFRA/water-abstraction-system/pull/15) we added a new service to start selecting charge versions to be included in the Supplementary Bill run.

But at the moment, its only filter is all charge versions with the `scheme` SROC. We have test data and scenarios to work with so we can now start implementing the filter 'for real'.

This is the start of the process. We expect it's likely we'll be finessing the filter in subsequent PRs. But this gets the ball rolling.

In this change, we are starting to filter the charge versions by

- are they linked to a licence that has been marked `include_in_supplementary_billing`
- are they 'current', which we have determined by whether `end_date` is populated (because `status` doesn't work)

** Notes

- Add migrations for 'water.licences' and 'water.charge_versions' tables

The `include_in_supplementary_billing` flag is in the `licences` table and not the `charge_versions` table. This means we can't just rely on checking the charge versions table and will need to build a licences table for our testing.

We also add a migration to add `licence_id` and `end_date` fields to the charge versions table. `licence_id` is how the 2 tables are linked and `end_date` is what we are having to rely on to determine the 'current' charge version.

- Expand the test data helper support

We add a new `LicenceHelper` as our query now needs these linked to charge versions. Rather than have our tests repeatedly set this up we have the charge version helper automatically call the licence helper and manage linking the 2 records. Adding the licence also means we need to start passing a bunch of data around now. So, we go with some defaults.

We thought we might need a `RegionHelper` so added that as well. Turns out we didn't, but we expect to in the future so left it in.

Co-authored-by: Jason Claxton <30830544+Jozzey@users.noreply.github.com>
Co-authored-by: Stuart Adair <43574728+StuAA78@users.noreply.github.com>