Add new SROC Billing Period service
https://eaflood.atlassian.net/browse/WATER-3802

One of the factors that determines whether a charge version is to be included in a bill run, is whether it matches the 'billing period'.

Initially, for SROC that's just going to be the current financial year. We're focused on getting supplementary billing sorted for this year but we know we'll need to iterate on this to handle multiple years in the future.

So, to limit the change when we do that we're adding a `BillingPeriodService` in this change which will calculate what financial years need to be checked for a bill run. It's just that as of now it will just return the current financial year.

Because we are adding this new service, we're also going to refactor out the current charge version query into its own service. `SupplementaryService` will then become the orchestrator for the whole process, and handle returning all relevant data to allow our QA team to confirm all is working as expected.