Move supplementary stuff out of test
We were putting things in `test/` because that is where we have our controller. But it's only the endpoint that is provided for testing, hence `/test`. The services we have been writing will be retained and used for real (in some form) in the future. So, it doesn't make sense to keep them in a `test/` folder. Plus, it's blummin confusing!

So, this quick change moves the services and their tests into `app/services/supplementary_billing`.