Refactor Supplementary output to test presenter
https://eaflood.atlassian.net/browse/WATER-3787
https://eaflood.atlassian.net/browse/WATER-3802
https://eaflood.atlassian.net/browse/WATER-3826

We're trying to make it easy to test that we are selecting the right values from the DB when generating the SROC supplementary bill run. Our test endpoint currently list the charge versions, which includes the licence reference. In [Add new SROC Billing Period service](https://github.com/DEFRA/water-abstraction-system/pull/30) we add the financial years.

But it would be easier to validate the licences that have been selected if the response included a unique list of them rather than having to searching through the charge version results. This gives us a prime opportunity to split out the presentation logic from what the service is doing, especially as it is likely to be discarded when we are happy the process is working.

So, in this change we move formatting the response to a new 'presenter'. As part of the change we'll add a unique list of the Licences being used.