Use table aliases in query
We are going to amend the query to return more fields. But those fields will clash if we don't specify which table they come from. This is likely to explode the query as we have to prefix every field name.

If you were writing raw SQL, you'd alias the table names to avoid this. Knex allows us to do the same so we take advantage of this feature.