Add new SROC Billing Period service (#30)
https://eaflood.atlassian.net/browse/WATER-3802

One of the factors that determine whether a charge version is to be included in a bill run, is whether it matches the 'billing period'.

Initially, for SROC that's just going to be the current financial year. We're focused on getting supplementary billing sorted for this year but we know we'll need to iterate on this to handle multiple years in the future.

So, to limit the change when we do that we're adding a `BillingPeriodService` in this change which will calculate what financial years need to be checked for a bill run. It's just that as of now it will just return the current financial year.

Because we are adding this new service, we're also going to refactor out the current charge version query into its own service. `SupplementaryService` will then become the orchestrator for the whole process, and handle returning all relevant data to allow our QA team to confirm all is working as expected.

**Notes**

- Move supplementary stuff out of test
  We were putting things in `test/` because that is where we have our controller. But it's only the endpoint that is provided for testing, hence `/test`. The services we have been writing will be retained and used for real (in some form) in the future. So, it doesn't make sense to keep them in a `test/` folder. Plus, it's blummin confusing!

Co-authored-by: Jason Claxton <30830544+Jozzey@users.noreply.github.com>