Update licences response in supplementary
https://eaflood.atlassian.net/browse/WATER-3787

As part of our first pass of getting the supplementary process kicked off and working we are just trying to ensure that the records we are working with match the acceptance criteria of the tickets the team have created.

The acceptance criteria for WATER-3787 is all licences flagged for supplementary billing for a selected region. At the moment the licence results we returned are generated by extracting the unique licence references from our charge versions results. This means we're only showing licences after an additional filter has been applied.

Currently this is only for testing purposes so there is no fear of what we are doing making the final version. But to better align with the ticket and its acceptance criteria, this change updates the supplementary service to return a seperate licences result.