Fix tests for new dev environment
https://github.com/DEFRA/water-abstraction-team/issues/64

Having built a new development environment based on Docker and Docker Compose that allows us to setup and maintain WRLS on Apple M1 Mac's we've hit some further issues. A number of the repos have it had coded into the tests to assume everything is always running on `http://localhost`. When it comes to our new Docker environment though, the Addressbase Facade and the Charge Module aren't anymore.

When it comes to this repo, the other assumption is that the machine the app is running on has the redis CLI installed. In our new dev environment that isn't the case, and shouldn't be assumed in our AWS environments.

So, this change

- fixes unit tests not to depend on hard coded urls
- makes the redis based tests pending (we'll tackle those properly once we have implemented redis support)