Amend the supplementary charge version query
https://eaflood.atlassian.net/browse/WATER-3826

The query as we have it is to simplistic. We need to review all charge versions within a given financial year to accurately calculate the bill run. We can't just look at the 'current' charge version.

So, this change updates the query to consider the creation and effective dates for the charge versions linked to a licence.