Fix tests for new dev environment (#40)
https://github.com/DEFRA/water-abstraction-team/issues/64

Having built a new development environment based on Docker and Docker Compose that allows us to set up and maintain WRLS on Apple M1 Mac's we've hit some further issues. A number of the repos have it had coded into the tests to assume everything is always running on `http://localhost`. When it comes to our new Docker environment though, the Addressbase Facade and the Charge Module aren't anymore.

So, this change updates the unit tests not to depend on hard-coded URLs.