Amend the supplementary charge version query (#41)
https://eaflood.atlassian.net/browse/WATER-3826

The query as we have it is too simplistic. We need to review all charge versions within a given financial year to accurately calculate the bill run. We can't just look at the 'current' charge version.

So, this change updates the query to consider the start date (effective date) for the charge versions linked to a licence in conjunction with the billing period.

** Note

We're still working on our understanding of what is required when generating a supplementary bill run. At the moment the functionality is more about confirming we're selecting the right things rather than getting on with generating a bill run.

So, what we have will not be the finished version.