Fix Objection model relations for schemas
https://eaflood.atlassian.net/browse/WATER-3829

Our current solution for working with schemas is to specify the schema name in the `tableName()` getter respone.

That has been working fine when just working in isolation. But our [first attempt](https://github.com/DEFRA/water-abstraction-system/pull/43) to generate a query that uses relations is erroring.

```text
Licence.relationMappings.charge-versions: join: either `from` or `to` must point to the owner model table and the other one to the related table. It might be that specified table 'water.charge_versions` is not correct
```

Some digging later we found [this issue](https://github.com/Vincit/objection.js/issues/85) which highlights our solution won't work

> Setting the schema to the tableName doesn't work. knex doesn't interpret the dot as a schema separator, but as a part of the table name.

So, this change implements the solution documented in that issue. We extend Objection's `QueryBuilder` with one that defaults all queries to use `withSchema()` taking the schema to use from a new model property `defaultSchema`.