Fix breaking test
Ok, so it was actually because we had not specified the table name properly. It all works if we correct that. So, no need to implement the solution defined in https://github.com/Vincit/objection.js/issues/85; knex seems to be handling fine the use of dot notation.

It might not _truly_ be using schema. But if we stick to this format it will give us the flexibility of making relationships across schemas. So, this is a mid-PR change in direction. Now we're just fixing things rather than implementing a new 'ways-of-working'.