Fix Objection model relations (#44)
https://eaflood.atlassian.net/browse/WATER-3829

Our [first attempt](https://github.com/DEFRA/water-abstraction-system/pull/43) to generate a query that uses relations is erroring.

```text
Licence.relationMappings.charge-versions: join: either `from` or `to` must point to the owner model table and the other one to the related table. It might be that specified table 'water.charge_versions` is not correct
```

Initially, we thought it was the way we're working with schemas as highlighted in [this issue](https://github.com/Vincit/objection.js/issues/85). But actually, it was just a typo.

So, this change fixes that. It also corrects some other details of the relationships and adds some tests just to ensure they are working as expected.