Update licence query to just return SROC (#43)
https://eaflood.atlassian.net/browse/WATER-3787

The original acceptance criteria for what licences to consider was anything flagged as 'include in supplementary billing'.

There are reasons why those that are void, deferred, revoked etc still need to be included. But our current scope is to generate SROC supplementary bill runs. So, on review, it makes sense that we only look at licences that have at least 1 SROC charge version.

This change updates our query to filter out anything that isn't SROC.