Replacing classes with modules and functions
> Classes are a template for creating objects.

That is the first sentance of the Mozilla documentation for **Classes**. It is what those of us from other languages understand classes to be for. Some of us though (we're looking at you @Cruikshanks) classes are _all things_!

Because of that thinking, we've allowed Classes to creap in where they shouldn't really be used. Most noticeably, things like controllers, presenters and services, which are formed of purely static methods. There is no real reason to wrap their functionality in a class. They could just be expressed as modules and export their relevant functions.

That's what those experienced with JavaScript are more likely to expect. By not doing things that way we're breaking the [Principle of least astonishment](https://en.wikipedia.org/wiki/Principle_of_least_astonishment).

We're believe classes are right for our Models, as it follows the convention of [Objection](https://vincit.github.io/objection.js/guide/models.html) and we do create instances of them. Also, if we have need to create objects in the future, classes will be our go-to.

But before this repo gets to far down the road we're taking this opportunity to refactor things from classes into modules.