Refactor Hapi-pino logger to clean things up (#49)
The first tweak is around config. We were pulling it into our main `app/server.js` and `HapiPinoIgnoreRequestService`. This meant we also needed to stub the config to get tests working for the `HapiPinoIgnoreRequestService`. By moving the config into the plugin we can use it where it's needed, and ensure it gets passed through as an arg. This means `LogConfig` is now only referenced in one place and we can remove the stubbing in our `HapiPinoIgnoreRequestService` tests because we're passing the critical option through.

The second was a mix of conventions. We were using a separate service to contain the logic of whether a request should be logged but a function in the plugin to cover logging in tests. By moving the logging in tests logic to a service we get consistency _and_ we can write some unit tests to cover it.