Fix test licences query
https://eaflood.atlassian.net/browse/WATER-3787

We've built a query to demonstrate what licences we are considering when building an SROC supplementary bill run. It's been highlighted that the query is returning duplicate results. A quick check found it's where we have a licence flagged for supplementary billing that has more than 1 SROC charge version. The inner join means we get a row per charge version.

This change fixes the issue, making our results distinct.