Add generic http request module
https://eaflood.atlassian.net/browse/WATER-3833

Our `/service-status` page makes `GET` requests to the other services in order to check if they are 'alive'. We now need to start making requests to other external services, for example the Charging Module API and AWS Cognito (to authenticate with the CHA API).

There are options we want pass to [got](https://github.com/sindresorhus/got) all the time. Plus we want a standard way of handling the response. **got**, by default, will throw an error for anything other than a `2xx` or `3xx` response. But we normally want the response, including status code and body as it will determine what we do next.

So, it's time to add a generic http request module to centralise and standardise these things.