Add `plainOutput` option to `ErrorPagesPlugin`
Our `ErrorPagesPlugin` will format Boom errors into friendly HTML error pages. We therefore started down the path of manually setting the http status code in our error response. On reflection, we would prefer to use Boom errors everywhere. We therefore add a `plainOutput` option to the plugin so we can override the formatting if we wish.