Add initial RequestLib module
We know we'll need to support at least `GET`, `POST` `PATCH` and `DELETE` requests via Got in the future. So, building this on our service pattern isn't ideal (we don't want `GetService`, `PostService` etc).

We do already have a pattern for actions within a context that do similar things; our notifier lib modules. So, we've opted to add this initial version of `RequestLib` with a `get()` action as a module under `app/lib`.