Change from `validateAsync` to just `validate`
We started using Joi's `validateAsync` method to perform validation, which returns the validated data or throws an error if validation fails. On reflection this would make error handling more complicated -- we'd need to wrap the validation call within a controller in its own try-catch block, so that we can ensure the right kind of Boom error is returned. We therefore change this out for the standard `validate` method, which returns an object containing the validated data (within the `value` key) and an `error` key with any validation errors.