Move /service-status to /health and tidy health
https://github.com/DEFRA/water-abstraction-team/issues/54

Bit of housekeeping. Adding a new /service-status page was one of our initial tasks with this repo, intended to help us understand just how we'd go about fitting this app into our existing service.

Now we're adding more and more services it's time to do a little tidying.

This moves the `/service-status` page to the `/health` path. We also move services used by the health pages into a `health/` folder.