Add `idColumn` static method to `BillingBatchModel`
As we start to look at creating entries in the db, we run into an issue when Objection errors because it can't find an `id` column. On investigation it turns out that if the primary id column isn't called `id`, it needs to be specified in the model. We therefore add an `idColumn` static method to our billing batch model