Fix issue with error-pages.plugin
In [Add Create Sroc Bill Run endpoint](https://github.com/DEFRA/water-abstraction-system/pull/51) we realised the `error-pages.plugin.js` had been written with the assumption we would always want to return a HTML response. But we're now building API only endpoints for use by the other apps in the service. For this we just want the standard response to be returned.

So, we made a tweak and added additional options to the routes of our API only endpoints that allow us to set whether error-pages should use 'plain output' (no HTML).

Only, we've inadvertently broken things for routes that don't have that additional option set! 🤦

This change adds some tests in the hope we catch this sooner in future. It also fixes the issue.