Fix issue with error-pages.plugin (#60)
In [Add Create Sroc Bill Run endpoint](https://github.com/DEFRA/water-abstraction-system/pull/51) we realised the `error-pages.plugin.js` had been written with the assumption we would always want to return an HTML response. But we're now building API-only endpoints for use by the other apps in the service. For this we want the standard response to be returned.

So, we tweaked and added additional options to the routes of our API-only endpoints that allow us to set whether error-pages should use 'plain output' (no HTML).

But in doing so we inadvertently broke things for routes that don't have that additional option set! 🤦

This change adds some tests in the hope we catch this sooner in future. It also fixes the issue.

```
Airbrake controller: GET /status/airbrake
  ✔ 19) returns a 500 error (6 ms)
Debug: internal, implementation, error 
    TypeError: Cannot read properties of undefined (reading 'plainOutput')
    at /home/repos/water-abstraction-system/app/plugins/error-pages.plugin.js:25:363
    at exports.Manager.execute (/home/repos/water-abstraction-system/node_modules/@hapi/hapi/lib/toolkit.js:57:29)
    at Request._invoke (/home/repos/water-abstraction-system/node_modules/@hapi/hapi/lib/request.js:397:55)
    at Request._postCycle (/home/repos/water-abstraction-system/node_modules/@hapi/hapi/lib/request.js:468:86)
    at Request._reply (/home/repos/water-abstraction-system/node_modules/@hapi/hapi/lib/request.js:447:20)
    at Request._execute (/home/repos/water-abstraction-system/node_modules/@hapi/hapi/lib/request.js:281:14)
```