Implement abstraction period query using Objection
https://eaflood.atlassian.net/browse/WATER-3862

We want to extract data on relevant charge versions, their references and abstraction periods from the DB. We’ll then use this to determine what extractions periods apply for a given charge version and billing period.

The first step is getting the data out of the DB. We need to implement that query using our ORM Objection.