Refactor CreateEventService for bill runs
https://eaflood.atlassian.net/browse/WATER-3831

We've created the model and a service wrapper for creating events. But at this time we are only interested in creating an `event` when a new bill run is initialised. That means some fields will always be set with a certain value.

Also, based on reviewing existing events created by the legacy code the `metadata` holds a formatted version of a bill run. So, we need to handle that.

All this means we'd be better placed to refactor the existing service to focus on new bill run events.